Removing Tenant from header and changes in Alert
Aerogear accepts URI with GET in pipe, but not with PUT or POST.
So Complete URL are used for an alert ack and resolve.

New method savePipe is for providing general POST and PUT.

TENANT is no longer needed to be specified in the header

Complete treatment for the list of triggers connected to resource still in progress.